+++ src/services/@l13/jsonc.test.ts
@@ -517,6 +517,70 @@ describe('jsonc', () => {
 			
 		});
 		
+		describe('remove trailing comma in objects', () => {
+			
+			runTests([
+				{
+					desc: 'trailing comma in singleline object',
+					expect: '{"a":1,}',
+					toBe: { a: 1 },
+				},
+				
+				{
+					desc: 'trailing comma in singleline object with space',
+					expect: '{"a":1, }',
+					toBe: { a: 1 },
+				},
+				
+				{
+					desc: 'trailing comma in singleline object with tab',
+					expect: '{"a":1,\t}',
+					toBe: { a: 1 },
+				},
+				
+				{
+					desc: 'trailing comma in multiline object',
+					expect: `{
+						"a":1,
+					}`,
+					toBe: { a: 1 },
+				},
+			]);
+			
+		});
+		
+		describe('remove trailing comma in arrays', () => {
+			
+			runTests([
+				{
+					desc: 'trailing comma in singleline array',
+					expect: '[1,]',
+					toBe: [1],
+				},
+				
+				{
+					desc: 'trailing comma in singleline array with space',
+					expect: '[1, ]',
+					toBe: [1],
+				},
+				
+				{
+					desc: 'trailing comma in singleline array with tab',
+					expect: '[1,\t]',
+					toBe: [1],
+				},
+				
+				{
+					desc: 'trailing comma in multiline array',
+					expect: `[
+						1,
+					]`,
+					toBe: [1],
+				},
+			]);
+			
+		});
+		
 	});
 	
 });